chore: extract common functions from metactl (#14698)
